feat: TAS-46 refactor newCmd
